Add support for register mavlink handler
