+++ src/module/mavproxy/mavproxy_monitor.h
@@ -1,0 +1,32 @@
+/******************************************************************************
+ * Copyright 2023 The Firmament Authors. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *****************************************************************************/
+
+#ifndef MAVPROXY_MONITOR_H__
+#define MAVPROXY_MONITOR_H__
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+fmt_err_t mavproxy_monitor_create(void);
+fmt_err_t mavproxy_monitor_register_handler(uint8_t chan, fmt_err_t (*handler)(mavlink_message_t*, mavlink_system_t));
+fmt_err_t mavproxy_monitor_deregister_handler(uint8_t chan, fmt_err_t (*handler)(mavlink_message_t*, mavlink_system_t));
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif
