+++ src/module/sysio/pilot_cmd.c
@@ -107,7 +107,7 @@ static void mavlink_send_pilot_cmd(void)
 
     mavlink_msg_fmt_pilot_cmd_encode(mav_sys.sysid, mav_sys.compid, &msg, &mav_pilot_cmd);
 
-    mavproxy_send_immediate_msg(MAVDEV_GCS_CHAN, &msg, false);
+    mavproxy_send_immediate_msg(MAVPROXY_GCS_CHAN, &msg, false);
 }
 #endif
